backend/pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = "API dashboard"
authors = ["Mikhail <podbereznikov34@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standart"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
